# server/openapi.yaml
openapi: 3.0.0
info:
  title: Book Management API
  version: 1.0.0
  description: A simple CRUD API for managing books

servers:
  - url: http://localhost:5000/api

paths:
  /api/books:
    post:
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    get:
      summary: Retrieve all books
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'

  /books/{id}:
    put:
      summary: Update a book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    delete:
      summary: Delete a book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Book deleted successfully

components:
  schemas:
    BookInput:
      type: object
      required:
        - title
        - author
      properties:
        title:
          type: string
          example: "Atomic Habits"
        author:
          type: string
          example: "James Clear"
        genre:
          type: string
          example: "Self-help"
        publishedYear:
          type: integer
          example: 2018

    Book:
      allOf:
        - $ref: '#/components/schemas/BookInput'
        - type: object
          properties:
            _id:
              type: string
              example: "64b5fc92cfb3b251d6f3e317"